refactor: Migrate orbit-mcp tests to per-module tests/ layout [ORB-00228]
